Refactoring Network Layer
