+++ app/src/main/java/com/unsa/suppliers/data/network/services/CountryService.kt
@@ -1,7 +1,6 @@
 package com.unsa.suppliers.data.network.services
 
-import com.unsa.suppliers.data.dtos.main.countries.CountryRequest
-import com.unsa.suppliers.data.dtos.main.countries.CountryResponse
+import com.unsa.suppliers.data.dtos.main.countries.*
 import com.unsa.suppliers.data.network.clients.CountryApiClient
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
